+++ Classes/Service/GlobalVariableProviders/RequestParameterProvider.php
@@ -11,12 +11,13 @@ declare(strict_types=1);
 namespace PSB\PsbFoundation\Service\GlobalVariableProviders;
 
 use JsonException;
+use PSB\PsbFoundation\Utility\ContextUtility;
 use PSB\PsbFoundation\Utility\StringUtility;
-use PSB\PsbFoundation\Utility\VariableUtility;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\NotFoundExceptionInterface;
 use TYPO3\CMS\Core\Utility\ArrayUtility;
 use TYPO3\CMS\Extbase\Configuration\Exception\InvalidConfigurationTypeException;
+use function is_array;
 
 /**
  * Class RequestParameterProvider
@@ -26,42 +27,6 @@ use TYPO3\CMS\Extbase\Configuration\Exception\InvalidConfigurationTypeException;
 class RequestParameterProvider extends AbstractProvider
 {
     /**
-     * @param string $key
-     * @param bool   $strict
-     *
-     * @return mixed
-     * @throws ContainerExceptionInterface
-     * @throws InvalidConfigurationTypeException
-     * @throws JsonException
-     * @throws NotFoundExceptionInterface
-     */
-    public static function getRequestParameter(string $key, bool $strict = false): mixed
-    {
-        return VariableUtility::getValueByPath(self::getRequestParameters(), $key, $strict);
-    }
-
-    /**
-     * @TODO: Remove public access. Data should be retrieved via GlobalVariableService only!
-     *
-     * @return array
-     * @throws ContainerExceptionInterface
-     * @throws InvalidConfigurationTypeException
-     * @throws JsonException
-     * @throws NotFoundExceptionInterface
-     */
-    public static function getRequestParameters(): array
-    {
-        $parameters = $_GET ?? [];
-        ArrayUtility::mergeRecursiveWithOverrule($parameters, $_POST ?? []);
-
-        array_walk_recursive($parameters, static function (&$item) {
-            $item = StringUtility::convertString($item);
-        });
-
-        return $parameters;
-    }
-
-    /**
      * @return array
      * @throws ContainerExceptionInterface
      * @throws InvalidConfigurationTypeException
@@ -70,6 +35,18 @@ class RequestParameterProvider extends AbstractProvider
      */
     public function getGlobalVariables(): array
     {
-        return self::getRequestParameters();
+        $request = ContextUtility::getRequest();
+        $parameters = $request?->getQueryParams();
+        $postParameters = $request?->getParsedBody();
+
+        if (is_array($postParameters)) {
+            ArrayUtility::mergeRecursiveWithOverrule($parameters, $postParameters);
+        }
+
+        array_walk_recursive($parameters, static function(&$item) {
+            $item = StringUtility::convertString($item);
+        });
+
+        return $parameters;
     }
 }